define more move functions and vops
The compiler needs to know how to move quantities between locations
and between representations.  The MOVE-VOPs are used to move them
within a routine; the MOVE-FUNs are for register spilling and
save/restore, at least according to a comment.

We also need to extend the storage base definition of ANY-REG, so that
it knows that constants in the IMMEDIATE storage class can be moved
into ANY-REG locations.

This gets us as far as failing to pack (choose actual registers for
our Temporary Names) which is reasonable since we haven't actually
given any of these storage classes any actual registers.