Strengthen the type of the META-INFO-DEFAULT slot.
By coercing the default to a function at meta-info construction time,
we save a type test when actually executing INFO in the more common
case of having a defaulting function. (In the less common case of
having a default object instead of lambda, we incur the cost of a
function call).