Be stricter about using the correct package for xc float ops.
Instead of canonicalizing the args automatically, use the normal
XC-STRICT-CL machinery to ensure no one tries to e.g. COERCE to a host
float.

Also remove unused floating point format flag from xc float math
file. It was never used, and if we did support multiple floating point
formats for the host, this flag doesn't matter for that case any more
since we use it only for cache and cross-check for a given build; it's
no longer version-controlled as an oracle.