Rework alien callables for redefinition and docstrings.
* Document and define redefinition semantics for alien callables,
allowing incremental redefinition. These semantics are based on how
CMU CL handles redefining callbacks, with the difference that type
incompatibility is stricter, and incompatible redefinitions invalidate
the old callback. I'm not sure whether have strict type equality or
having a weaker notion of compatibility is better. The docstring is
cribbed from CMU CL, and the interface now matches their DEF-CALLBACK
closely otherwise.

* Allow attaching docstrings in SB-ALIEN:DEFINE-ALIEN-CALLABLE.

* The SB-ALIEN:DEFINE-ALIEN-CALLABLE interface does linkage by name,
so we can remove various non-exported functions that tried to deal in
ALIENs directly, and we don't need to bother with caching callback
SAPs by function.

* ALIEN-LAMBDA now returns a Lisp function rather than a callback sap
to support redefinition semantics better. The symbol is unexported but
is used by CFFI, which is addressed in this pull request
https://github.com/cffi/cffi/pull/406.