Auto-stack-allocate local functions even with multiple references.
With all the machinery in place to insert cleanup code correctly given
multiple derived dynamic extents, we just need to weaken the condition
for stack allocating a closure from having a single reference to
having all references be downward funarg uses.

Update the manual not only to document this new optimization but also
the existing way for users to declare downward funargs for their own
functions.

Also, there are a couple cases in the self-build where this
optimization applies. Remove those explicit dynamic extent
declarations which no longer do (or never did) anything and let the
compiler prove/auto-stack-allocate many of the valid cases. This makes
it easier to find the cases that the compiler cannot prove safe to
stack allocate yet, which led to finding some potential optimizations
that don't work yet, which failing test cases have been added for.