stack: Preserve intervening stack allocated objects differently.
Instead of keeping an explicit preserve set when we detect an
intervening dynamic extent, we can just set the stack lvar of the
inteervening dynamic extent to be that of the longer dynamic
extent. This is simpler and also reduces the number of stack lvars
floating around which just get cleaned up at the same time.

This fixes a bug with a SETQ on a dynamic extent variable nested on
top of a stack allocation with shorter extent inside a loop.